+++ example/book/intelligence/character/bots/character_description_llm_bot.py
@@ -7,6 +7,7 @@ from typing_extensions import TYPE_CHECKING
 from dandy.llm import BaseLlmBot, Prompt
 from example.book.intelligence.character.enums import CharacterType
 from example.book.intelligence.character.intel import CharacterIntel
+from example.book.intelligence.character.prompts import characters_intel_prompt
 
 if TYPE_CHECKING:
     from example.book.intelligence.intel import BookIntel
@@ -15,6 +16,10 @@ if TYPE_CHECKING:
 
 class CharacterGeneratorLlmBot(BaseLlmBot):
     config = 'ADVANCED'
+    instructions_prompt = (
+        Prompt()
+        .text('You are a character generating bot, please create a character based on the provided input.')
+    )
     
     @classmethod
     def process(
@@ -25,8 +30,6 @@ class CharacterGeneratorLlmBot(BaseLlmBot):
     ) -> CharacterIntel:
         prompt = Prompt()
         
-        prompt.text(f'Generate a {character_type.value} Character for the following book.')
-        
         prompt.line_break()
         
         prompt.text(f'Title: {book_intel.start.title}')
@@ -35,17 +38,7 @@ class CharacterGeneratorLlmBot(BaseLlmBot):
         if characters_intel:
             prompt.line_break()
             
-            prompt.text('Other Characters:')
-            
-            for character in characters_intel.characters:
-                prompt.line_break()
-                
-                prompt.text(f'Name: {character.first_name} {character.last_name}')
-                prompt.text(f'Age: {character.age}')
-                prompt.text(f'Nickname: {character.nickname}')
-                prompt.text(f'Description: {character.description}')
-                prompt.text(f'Type: {character.type.value}')
-                prompt.text(f'Alignment: {character.alignment.value}')
+            prompt.prompt(characters_intel_prompt(characters_intel))
     
         return cls.process_prompt_to_intel(
             prompt=prompt,