+++ example/book/intelligence/intel.py
@@ -1,6 +1,7 @@
 from typing import Union
 
 from dandy.intel import BaseIntel
+from example.book.intelligence.chapter.intel import ChaptersIntel
 
 from example.book.intelligence.character.intel import CharactersIntel
 from example.book.intelligence.plot.intel import PlotIntel
@@ -17,4 +18,5 @@ class BookIntel(BaseIntel):
     characters: Union[CharactersIntel, None] = None
     plot: Union[PlotIntel, None] = None
     world: Union[WorldIntel, None] = None
+    chapters: Union[ChaptersIntel, None] = None