+++ example/book/intelligence/plot/bots/plot_outline_llm_bot.py
@@ -4,7 +4,7 @@ from typing_extensions import TYPE_CHECKING
 
 from dandy.llm import BaseLlmBot, Prompt
 from example.book.intelligence.plot.intel import PlotIntel
-
+from example.book.intelligence.prompts import book_intel_prompt
 
 if TYPE_CHECKING:
     from example.book.intelligence.intel import BookIntel
@@ -16,7 +16,7 @@ class PlotOutlineLlmBot(BaseLlmBot):
         Prompt()
         .text('You are a plot bot. You will be given a book title, overview, world and characters.')
         .text('You will generate a plot using the classic heroes journey plot structure.')
-        .text('Create 10 plot points with only an outline and no description.')
+        .text('Create 10 plot points with only an outline (do not number them) and no description.')
     )
    
     @classmethod    
@@ -24,5 +24,8 @@ class PlotOutlineLlmBot(BaseLlmBot):
             cls,
             book_intel: BookIntel
     ) -> PlotIntel:
+        return cls.process_prompt_to_intel(
+            prompt=book_intel_prompt(book_intel),
+            intel_class=PlotIntel
+        )
         
-        pass